Condition mention of BeforeGenerateProjectPriFile only if needed
If you have a target that mentions another target in BeforeTargets,
but your target won't run, MSBuild still gives a warning. The workaround
is to condition the computation of BeforeTargets as well. There's a
bigger question here about where to move the .NET Core injection to,
since it's ugly that it lives here; I ignore that question for now.

Fixes NuGet/Home#1814.