Pass references from projects to ResolveAssemblyReferences.AssemblyFiles
If we pass a full path to ResolveAssemblyReferences.Assemblies, it will
remove the file and not pass it along to the compiler. This is bad
for project-to-project references, since the Visual Studio IDE still
needs to know about them to make sure everything is wired up correctly.
With this, we now pass project-to-project references to AssemblyFiles,
matching the behavior of the ResolveProjectReferences target.

There is a deeper question of whether or not we should pass all our
full paths along to AssemblyFiles, but that's something which requires
fairly extensive testing and might have back-compat problems that need
to be dealt with.