Support reading package folder locations out of the lock file
The task previously was constructing the path to the user's packages
location itself, which had various issues. It meant that if the restore
was done with a different path, we don't know that. It also meant we
didn't handle the paths specified in nuget.config. Now, if the paths
are present in the lock file, we'll use that. Otherwise, we'll construct
our old path as before. Any explicitly given path to the task is
still used no matter what.