Avoid stale project.assets.json files
Imagine you have a project with a single `PackageReference` item. This will cause us to generate a project.assets.json file in the intermediates directory. Then you remove that item, and run NuGet restore.

That *should* cause up to drop any references to the old item, but currently it does not. Since there are no `PackageReference` items we don't regenerate the projec.assets.json, leaving the stale one in the intermediates directory. And since we can still find it, we still use it.

The basic idea behind the fix is that, just as we should only use a project.lock.json file if and only if there is a project.json file, we should only use a project.assets.json file if and only if the project contains `PackageReference` items.

This is simple to describe, but it turns out to be relatively difficult to implement in MSBuild. We can only reliably check for the existence of items with a particular target, which means our current up-front logic to decide between project.lock.json, project.assets.json, or neither has to go. Instead, we calculate both the path to project.lock.json and project.assets.json upfront, and only within the targets that need them do we decide which to use.

In addition, any place that checks for the existence of project.lock.json must now check for both `PackageReference` items and the existence of project.assets.json as well. This leads to some unfortunately (but unavoidably) complicated `Condition` attributes.