Ignore case when matching package names
Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/500532.

We may need to vary which packages (or versions of packages) to pull in depending on the target framework. In the assets/lock file this is supported by the projectFileDependencyGroups section. This contains an entry for each supported framework, and the entry lists the packages used when targeting that framework. In addition, there is a "universal" entry (denoted by an empty name) for packages that are used in all frameworks.

To determine which packages to list in Solution Explorer we match items in the relevant projectFileDependencyGroups item against the packages listed in the "libraries" section. However, we are currently doing this comparison in a case-sensitive manner, even though NuGet packages names are case-insensitive. This means that if the package name differs in case between the two, we won't find it and it won't show up in Solution Explorer. This is "just" a display issue, as the build and Intellisense are unaffected. Still, it is disconcerting for the user, as packages may suddenly "disappear" during package management operations.