Disable ResolveNuGetPackages for *.vcxproj
I have determined that the ResolveNuGetPackageAssets task
does not function under a *.vcxproj file, no matter what I
try. Inspection of the task code reveals no indications of
why it is failing. However, since this task is aimed towards
restoring .NET assets, and that NuGet doesn't support .NET
assets in C++/CLI projects anyway, this is not much of a loss.

Import of props/targets files in build/native folders of packages
is not impacted by this change.